requirements.txt
geoip2
fire	
elasticsearch==7.10.1
python-dotenv